tangle: Use bash instead of zsh
The default base install of many major Linux systems doesn't include
zsh, causing this script to break on those.

Since it uses no zsh-specific features, it runs just as well with
bash, and that's part of the base install of all non-embedded and
non-pathological Linux distributions. So switch it to that.