INCOMPLETE: tangle: Do stack setup and build before we try to run tangle
Because of course in a fresh checkout, `pandoc-tangle` doesn't exist
because it has never been built. (We'll still need to do another build
after building it because the literate source may be updated and
change the compiler source whence we build `pandoc-tangle`.

XXX This still doesn't work, but we're closer.